Server/requirements.txt
flask
paho-mqtt
prometheus-client
